# packages/contracts/openapi/dabboba.openapi.yaml
openapi: 3.1.0
info:
  title: DABBOBA API
  version: 0.1.0
  description: Shared user, commerce, and privileged operations contract. Fixture-only prototype behavior is not represented as production persistence.
servers:
  - url: http://127.0.0.1:8788
    description: Isolated local API
...
paths:
...
  /v1/auth/exchange:
    post:
      tags: [Auth]
      operationId: exchangeCustomerAccessToken
      description: Verifies an asymmetric Supabase access token for exactly one Kakao, Naver, or phone identity and returns an opaque DABBOBA session. Accounts are keyed only by the broker issuer and subject, never email or phone; multi-provider broker identities require a separate explicit linking flow and are rejected here.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerAccessTokenExchangeInput" }
      responses:
        "201":
          description: DABBOBA customer session created.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503":
          description: Supabase customer token exchange is not configured.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
...
  responses:
    BadRequest:
      description: Request input is malformed or references an inactive catalog item.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Authentication is required or the session is invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: The authenticated actor lacks the required role or ownership.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    Conflict:
      description: The requested state transition conflicts with committed state.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    RateLimited:
      description: Too many attempts.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    CustomerAccessTokenExchangeInput:
      type: object
      additionalProperties: false
      required: [accessToken]
      properties:
        accessToken: { type: string, minLength: 64, maxLength: 16384, writeOnly: true }
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object, additionalProperties: true }
...
    UserRole: { type: string, enum: [USER, ADMIN, SUPER_ADMIN] }
    UserStatus: { type: string, enum: [ACTIVE, SUSPENDED, BANNED, DELETED] }
...
    Actor:
      type: object
      additionalProperties: false
      required: [userId, email, nickname, role, status, sessionId, permissions]
      properties:
        userId: { type: string, format: uuid }
        email: { type: [string, "null"], format: email }
        nickname: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        status: { $ref: "#/components/schemas/UserStatus" }
        sessionId: { type: string, format: uuid }
        permissions: { type: array, items: { type: string } }
...
    SessionCreated:
      type: object
      additionalProperties: false
      required: [token, expiresAt, actor]
      properties:
        token: { type: string }
        expiresAt: { type: string, format: date-time }
        actor: { $ref: "#/components/schemas/Actor" }
...